requirements.txt
langchain==0.3.27
langchain-openai==0.3.31
langgraph==0.4.9
pydantic==2.11.7
PyYAML==6.0.2
